+++ packages/quill/rollup.config.js
@@ -1,0 +1,10 @@
+import rollupConfig from '../../rollup.config.js'
+export default {
+  ...rollupConfig,
+  external: [
+    ...rollupConfig.external,
+    'react-quill',
+    '@react-corn/core',
+    'styled-components',
+  ],
+}